Update and rename assessmentObject.js to assessment.js
The Assessment really is considered a class and you dont want to add an object to the class name. When you instantiate a class such as on line 309 " var a = new Assessment();" a is now an object of class Assessment. Since those instances for the 3 objects are only used in a few places its probably ok to have single letter variables but I would recommend having meaningful names.

I also proposed an enhancement where you have a dictionary to keep track of all the checkboxes and added comments and an example in the code